+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## master
 
+- added support for single panel and sticky tooltip functionality
+- displaying a sticky tooltip now disables all other tooltips until the sticky tooltip is closed
 - removed auto positioning (did not like implementation and not really sure if it is needed)
-- tooltip content does not have to be with-in the main tooltip element
 - changed how to define tooltip positioning
 - a tooltip can now have both a regular hover tooltip content and a click sticky tooltip content